JxlEncoderFrameSettingsSetOption to take int64_t (#1617)
This change widens the type of the `value` argument
of `JxlEncoderFrameSettingsSetOption`, from
`int32_t` to `int64_t`. This allows setting the
center_x and center_y arguments.

+++ lib/include/jxl/encode.h
@@ -911,7 +911,7 @@ JXL_EXPORT JxlEncoderStatus JxlEncoderSetExtraChannelName(JxlEncoder* enc,
  */
 JXL_EXPORT JxlEncoderStatus JxlEncoderFrameSettingsSetOption(
     JxlEncoderFrameSettings* frame_settings, JxlEncoderFrameSettingId option,
-    int32_t value);
+    int64_t value);
 
 /** Forces the encoder to use the box-based container format (BMFF) even
  * when not necessary.
